requirements.txt
python-telegram-bot>=20.0
yt-dlp>=2023.3.4
firebase-admin>=6.0.0
python-dotenv>=1.0.0
requests>=2.28.0
Pillow>=9.0.0
asyncio
aiohttp>=3.8.0
python-dateutil>=2.8.0
pytz>=2022.1
schedule>=1.1.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
ffmpeg-python>=0.2.0